requirements.txt
python-telegram-bot==13.7
requests==2.26.0
gunicorn
